+++ apps/health_campaign_field_worker_app/lib/pages/authenticated.dart
@@ -9,6 +9,7 @@ import 'package:flutter_portal/flutter_portal.dart';
 import 'package:isar/isar.dart';
 import 'package:location/location.dart';
 import 'package:registration_delivery/registration_delivery.dart';
+import 'package:survey_form/survey_form.dart';
 
 import '../blocs/localization/app_localization.dart';
 import '../blocs/projects_beneficiary_downsync/project_beneficiaries_downsync.dart';