+++ apps/health_campaign_field_worker_app/lib/utils/extensions/extensions.dart
@@ -1,7 +1,9 @@
-// Importing necessary packages and files
+import 'package:attendance_management/attendance_management.dart';
 import 'package:closed_household/utils/utils.dart';
 import 'package:collection/collection.dart';
+import 'package:complaints/utils/utils.dart';
 import 'package:digit_data_model/data_model.dart';
+import 'package:digit_location_tracker/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
@@ -9,6 +11,8 @@ import 'package:inventory_management/utils/utils.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:referral_reconciliation/utils/utils.dart';
 import 'package:registration_delivery/registration_delivery.dart';
+// Importing necessary packages and files
+import 'package:survey_form/utils/utils.dart';
 
 import '../../blocs/auth/auth.dart';
 import '../../blocs/project/project.dart';