+++ packages/attendance_management/lib/router/attendance_router.dart
@@ -9,9 +9,21 @@ class AttendanceRoute extends $AttendanceRoute {
 
   @override
   List<AutoRoute> routes = [
-    AutoRoute(page: ManageAttendanceRoute.page),
-    AutoRoute(page: AttendanceDateSessionSelectionRoute.page,),
-    AutoRoute(page: MarkAttendanceRoute.page,),
-    AutoRoute(page: AttendanceAcknowledgementRoute.page),
+    AutoRoute(
+      page: MarkAttendanceRoute.page,
+      path: 'mark-attendance',
+    ),
+    AutoRoute(
+        page: ManageAttendanceRoute.page,
+        path: 'manage-attendance',
+    ),
+    AutoRoute(
+      page: AttendanceDateSessionSelectionRoute.page,
+      path: 'attendance-date-session-selection',
+    ),
+    AutoRoute(
+      page: AttendanceAcknowledgementRoute.page,
+      path: 'attendance-acknowledgement',
+    ),
   ];
 }