+++ packages/registration_delivery/CHANGELOG.md
@@ -1,6 +1,18 @@
+## 1.0.3+4
+* Delivery action label and administered_success status update based on the project
+
+## 1.0.3+3
+* Updated Service model additionalDetails objectType in digit_data_model package
+
 ## 1.0.3+2
 
-* Delivery action label and administered_success status update based on the project
+* Bug Fix -
+  * fixed wrong status map of beneficiaryRefused in utils
+* Sorting tasks based on createTime
+
+## 1.0.4-dev.1
+
+* Integrated survey_form package
 
 ## 1.0.3+1